rename /help to /ezpr help & /ezpr to /ezpr submit
