+++ src/cmd/index.ts
@@ -1,4 +1,4 @@
-export * from "../ezpr/cmd";
-export * from "../help/cmd";
+export * from "../ezpr submit/cmd";
+export * from "../ezpr help/cmd";
 export * from "./getNameByID";
 export * from "./openModal";
